requirements.txt
fpdf==1.7.2
numpy==1.19.0
Pillow==7.2.0
scipy==1.5.1
